Some class and namespace renaming
