+++ mecanum_drive_controller/include/mecanum_drive_controller/speed_limiter.h
@@ -39,7 +39,7 @@
 #ifndef SPEED_LIMITER_H
 #define SPEED_LIMITER_H
 
-namespace diff_drive_controller
+namespace mecanum_drive_controller
 {
 
   class SpeedLimiter
@@ -100,6 +100,6 @@ namespace diff_drive_controller
     double max_acceleration;
   };
 
-} // namespace diff_drive_controller
+} // namespace mecanum_drive_controller
 
 #endif // SPEED_LIMITER_H
